[#137057261] Clinical activity is not too early without stop_datetime
Patient "MipsExclusion Early" is not expected to have an exclusion
reason because his Hospice Service clinical activity is too early.
However, without a stop_datetime, this clinical activity's interval
does not have an upper boundary and therefore overlaps with
the measurement period interval. To make this clinical activity
too early, it needs to have a stop_datetime that is less than
the start of the measurement period.