remove relevantDateTime
While assessments (and some other activities) have a concept of "relevantDateTime" (single datetime), the Able QDM variant has "relevantPeriod" (timespan). To help reduce confusion for CQL->Drools developers we've removed "relevantDateTime" and replaced it with "start of RelevantPeriod". This is a deviation from the QDM spec, so if this is revisisted later, realize that assessments will likely not have a relevantPeriod and this will need to be revisited.